+++ changelog.md
@@ -77,6 +77,7 @@
 * better disambiguation-page detection in english
 * remove wikitext from caption titles
 * support 3-level templates (whew!)
+
 ## 5.0.0
 * new `Table` class and `List` classes
 * improved table-parser - generate name `col1` instead of `col-0`
@@ -95,9 +96,10 @@
 * keep `#` anchor data in .links()
 * show links default-on in latex output, like in md and html
 * render html/latex/json 'soft redirect', instead of blank pages
-## 6.0.0
+
+## 6.0.0 🚨
 * support `.paragraphs()`
-* :warning: changes to output of `.json()`. cleaning-up redundant data. (sorry for the frequent changes here). :warning:
+* :warning: major changes to output of `.json()`. cleaning-up redundant data.:warning:
 * * remove top-level `templates` data (found in `section`) - resume it with `{templates:true}`
 * * remove top-level `citations` data (found in `section`) - resume it with `{citations:true}`
 * return empty arrays in `.json()` again  ¯\_(:/)_ /¯
@@ -112,3 +114,4 @@
 * include `interwiki()` results in `.links()`
 * support `follow_redirects` option to fetch
 * hide object data in console.logs
+* move ALL image urls from `upload.wikimedia.org/wikipedia/commons` to `wikipedia.org/wiki/Special:Redirect/file/` via #86
